Fix broken login
Redux devtools extension was improperly installed in `store.js`. Fix
properly returns a function as arg. White screen of death on attempt to
load site is fixed.

Login would hang on a loading screen. Login fetch was originally
abstracted into actions.js file. The login fetch was not completing
before the app's attempt to hydrate. Login fetch was moved back to the
login component and `fetchData` is now called on a successful response.